requirements.txt
Flask==1.1.1
flask-restx==0.5.0
Jinja2==3.0.3
itsdangerous==2.0.1
gunicorn==20.0.2
werkzeug==2.0.2

api>=0.0.7
# Bio>=1.3.3
behave==1.2.6
biopython==1.75
inflect==6.0.2
editdistance>=0.6.0
openpyxl==3.0.10
protobuf==3.20.0
py-ard==0.8.0
py2neo==2021.0.1
PyHamcrest==1.9.0
regex==2022.3.2
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3